refactor: reorganize test infrastructure and clean up verbose logging
- Split monolithic rest_integration_test.go into organized test modules by functionality
- Remove verbose logging statements from config and credentials services for cleaner output
- Reorder migration timestamp to maintain proper chronological sequence
- Add storage/ directory to gitignore for local development files
